+++ src/main/kotlin/com/soom/monolithic_api/domain/account/signup/controller/SignupController.kt
@@ -1,10 +1,10 @@
 package com.soom.monolithic_api.domain.account.signup.controller
 
 import com.soom.monolithic_api.domain.account.common.data.dto.AccountDto
-import com.soom.monolithic_api.domain.account.signup.dto.*
-import com.soom.monolithic_api.domain.account.signup.request.SignupRequest
-import com.soom.monolithic_api.domain.account.signup.request.StudentSignupRequest
-import com.soom.monolithic_api.domain.account.signup.request.TeacherSignupRequest
+import com.soom.monolithic_api.domain.account.signup.data.dto.*
+import com.soom.monolithic_api.domain.account.signup.data.request.SignupRequest
+import com.soom.monolithic_api.domain.account.signup.data.request.StudentSignupRequest
+import com.soom.monolithic_api.domain.account.signup.data.request.TeacherSignupRequest
 import com.soom.monolithic_api.domain.account.signup.service.EmailTokenDecodeService
 import com.soom.monolithic_api.domain.account.signup.service.SignupService
 import org.springframework.http.ResponseEntity
@@ -32,13 +32,13 @@ class SignupController(
 
     //학생 회원가입 요청을 DTO 로 치환한다.
     private fun requestToDto(request: StudentSignupRequest): StudentSignupDto = requestToDtoTemplate(
-        {pair: Pair<StudentSignupRequest, String> -> SigninStudentAdditionalDto(pair.first.classNumber, pair.first.admissionAt, pair.first.department)},
-        {auth, profile, student: SigninStudentAdditionalDto -> StudentSignupDto(auth, profile, student)}
+        {pair: Pair<StudentSignupRequest, String> -> SigninStudentAdditionalDto(pair.first.classNumber, pair.first.admissionAt, pair.first.department) },
+        {auth, profile, student: SigninStudentAdditionalDto -> StudentSignupDto(auth, profile, student) }
     ).invoke(request)
     //교사 회원가입 요청을 DTO 로 치환한다.
     private fun requestToDto(request: TeacherSignupRequest): TeacherSignupDto = requestToDtoTemplate(
-        {pair: Pair<TeacherSignupRequest, String> -> SigninTeacherAdditionalDto(pair.first.major, pair.first.teacherType)},
-        {auth, profile, teacher: SigninTeacherAdditionalDto -> TeacherSignupDto(auth, profile, teacher)}
+        {pair: Pair<TeacherSignupRequest, String> -> SigninTeacherAdditionalDto(pair.first.major, pair.first.teacherType) },
+        {auth, profile, teacher: SigninTeacherAdditionalDto -> TeacherSignupDto(auth, profile, teacher) }
     ).invoke(request)
     //회원가입 요청을 DTO 로 치환하는 템플릿 함수
     private fun <T: SignupRequest, T2: SignupAdditionalComponents, R: SignupDto> requestToDtoTemplate(
