+++ src/main/kotlin/com/soom/monolithic_api/domain/account/signup/data/dto/SignupDto.kt
@@ -1,4 +1,4 @@
-package com.soom.monolithic_api.domain.account.signup.dto
+package com.soom.monolithic_api.domain.account.signup.data.dto
 
 sealed class SignupDto(
     open val auth: SigninCommonAuthDto,
