+++ src/main/kotlin/com/soom/monolithic_api/domain/account/signup/data/type/SchoolEmailType.kt
@@ -1,4 +1,4 @@
-package com.soom.monolithic_api.domain.account.signup.type
+package com.soom.monolithic_api.domain.account.signup.data.type
 
 import com.soom.monolithic_api.domain.account.common.data.type.SchoolType
 
@@ -8,16 +8,20 @@ enum class SchoolEmailType(
 ) {
     광주_소프트웨어_마이스터고등학교(
         SchoolType.광주_소프트웨어_마이스터_고등학교,
-        regexValidator("s[0-9]@gsm.hs.kr")),
+        regexValidator("s[0-9]@gsm.hs.kr")
+    ),
     대덕_소프트웨어_마이스터고등학교(
         SchoolType.대덕_소프트웨어_마이스터_고등학교,
-        regexValidator(".*@dsm.hs.kr")),
+        regexValidator(".*@dsm.hs.kr")
+    ),
     대구_소프트웨어_마이스터고등학교(
         SchoolType.대구_소프트웨어_마이스터_고등학교,
-        regexValidator(".*@dgsm.hs.kr")),
+        regexValidator(".*@dgsm.hs.kr")
+    ),
     부산_소프트웨어_마이스터고등학교(
         SchoolType.부산_소프트웨어_마이스터_고등학교,
-        regexValidator(".*@bsm.hs.kr"))
+        regexValidator(".*@bsm.hs.kr")
+    )
 }
 
 fun regexValidator(regex: String): (String) -> Boolean = {
