+++ src/main/kotlin/com/soom/monolithic_api/domain/account/signup/service/SignupServiceImpl.kt
@@ -7,11 +7,11 @@ import com.soom.monolithic_api.domain.account.common.data.entity.TeacherEntity
 import com.soom.monolithic_api.domain.account.common.data.type.RoleType
 import com.soom.monolithic_api.domain.account.common.data.type.SchoolType
 import com.soom.monolithic_api.domain.account.common.repository.AccountRepository
-import com.soom.monolithic_api.domain.account.signup.dto.SignupDto
-import com.soom.monolithic_api.domain.account.signup.dto.StudentSignupDto
-import com.soom.monolithic_api.domain.account.signup.dto.TeacherSignupDto
+import com.soom.monolithic_api.domain.account.signup.data.dto.SignupDto
+import com.soom.monolithic_api.domain.account.signup.data.dto.StudentSignupDto
+import com.soom.monolithic_api.domain.account.signup.data.dto.TeacherSignupDto
 import com.soom.monolithic_api.domain.account.signup.exception.WrongSchoolEmailException
-import com.soom.monolithic_api.domain.account.signup.type.SchoolEmailType
+import com.soom.monolithic_api.domain.account.signup.data.type.SchoolEmailType
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 
@@ -22,7 +22,7 @@ class SignupServiceImpl(
     ): SignupService {
     //학생 회원가입 수행
     override fun signupStudent(dto: StudentSignupDto): AccountDto =
-        signup {signupDto: StudentSignupDto, pair->
+        signup { signupDto: StudentSignupDto, pair->
             StudentEntity(
                 0L, signupDto.profile.name, signupDto.profile.gender, signupDto.profile.birth,
                 signupDto.auth.email, pair.first,
@@ -31,13 +31,14 @@ class SignupServiceImpl(
         }.invoke(dto, passwordEncoder, accountRepository)
     //교사 회원가입 수행
     override fun signupTeacher(dto: TeacherSignupDto): AccountDto =
-        signup {signupDto: TeacherSignupDto, pair ->
+        signup { signupDto: TeacherSignupDto, pair ->
             TeacherEntity(
                 0L, signupDto.profile.name, signupDto.profile.gender, signupDto.profile.birth,
                 signupDto.auth.email, pair.first,
                 RoleType.OPERATOR, pair.second, "",
                 signupDto.teacher.major, signupDto.teacher.teacherType)
         }.invoke(dto, passwordEncoder, accountRepository)
+
     //회원가입 수행
     private fun <T: SignupDto, T2: AccountEntity> signup(makeEntity: (T, Pair<String, SchoolType>) -> T2)
     : (T, PasswordEncoder, AccountRepository) -> AccountDto = { dto, passwordEncoder, accountRepository ->
